% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmental_framework_single.R
\name{compartmental_framework_single}
\alias{compartmental_framework_single}
\title{compartmental_framework_single function}
\usage{
compartmental_framework_single(birth_rate, mortality_rate, constant = TRUE,
  Stock = 1, age = 85)
}
\arguments{
\item{birth_rate}{a parameter for model, can be a number or a range of numbers, must NOT be a negative value.}

\item{mortality_rate}{a parameter for model, can be a number or a range of numbers, must NOT be a negative value.}

\item{constant}{a boolean parameter indicating whether the input parameters are a range of numbers or just one single number.}

\item{Stock}{an initial value for the compartmental model.}

\item{age}{specifying the extend of x-axis.}
}
\value{
A gridExtra object, and also print out the compartmental model in ggplot style.
}
\description{
This function estimates a single compartmental system with only Stock, birth_rate (flowing in rate) and 
mortality_rate (flowing out rate). This function is sort of a toy example for setting up the background of the package.
}
\examples{
Example 1: (taking only one number for parameters)
compartmental_framework_single(birth_rate=0.01,
                               mortality_rate=0.015,
                               Stock=1,
                               age=85)



Example 2: (taking a range of numbers for parameters) 
compartmental_framework_single(birth_rate=seq(0.05,0.01,length.out = length(seq(1,85,by=0.5))),
                              mortality_rate=seq(0.01,0.015,length.out = length(seq(1,85,by=0.5))),
                              constant=FALSE,
                              Stock=1,
                              age=85)


}
\keyword{compartmental}
\keyword{framework}
