% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismod_v2.R
\name{dismod_v2}
\alias{dismod_v2}
\title{dismod_v2 function}
\usage{
dismod_v2(pops, alldeath_rate, condition_death_rate, condition_incidence_rate,
  duration, Stock = 10000, Condition = 1, start_age = 0,
  age_range = c(0, 1, seq(5, 85, by = 5)), end_age = 85, plots = TRUE)
}
\arguments{
\item{pops}{taking in a vector of age-gender-specific population, preferrably for a certain time (for example, in a year).}

\item{alldeath_rate}{taking in a vector of age-gender-specific all cause mortality rate, preferrably for a certain time (for example, in a year).}

\item{condition_death_rate}{taking in a vector of age-gender-specific mortality rate for a specific condition (a disease), preferrably for a certain time (for example, in a year).}

\item{condition_incidence_rate}{taking in a vector of age-gender-specific incidence rate for a specific condition (a disease), preferrably for a certain time (for example, in a year).}

\item{duration}{a vector specifying the age-gender-specific duration of the disease.}

\item{Stock}{an initial value for the compartmental model (for Stock population--susceptible population).}

\item{Condition}{an initial value for the compartmental model (for diseased population--population with condition).}

\item{start_age}{specifying the lower bound of x-axis.}

\item{age_range}{manually setting up the age label vector. In real-world data, mortality and incidence data are often specified in this way: "0", "1-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39", "40-44", "45-49", "50-54", "55-59", "60-64", "65-69", "70-74", "75-79", "80-84", "85+". The default value for this parameter is to mimic the real world situation.}

\item{end_age}{specifying the upper bound of x-axis.}

\item{plots}{a boolean argument, specifying whether to visualize the compartmental framework model or not. The plots are in a ggplot fashsion.}
}
\value{
A gridExtra object, when plots=TRUE;

or a list of data containing the simulated data produced by differentiation in the modeling process, when plots=FALSE. The return data also contains prevalence estimated from the input incidence and mortality.
}
\description{
This function simulates the process behind DisMod-II (developed by WHO, available at www.who.int/evidence/dismod, for windows-PC only) and DisMod-MR (developed by IHME). Function in current forms only consider Stock and Condition (which is used in DisMod-MR), leaving out the Dead group (which is used in DisMod-II).
}
\details{
Beware that currently, this function is like a toy example, only taking in incidence data and mortality data
as its main input data. The reason to do so is that the author currently only came across situation where
incidence and mortality data are available (incidence and mortality data are already readily available in the death registry system).

The function in current forms only take in pops, alldeath_rate, condition_death_rate, condition_incidence_rate and duration where these arguments have an exact match in vector length with each other.

Also, this model calculates remission with the equation: remission = 1/duration - fatality (adapted from DisMod-II). fatality is computed as: fatality = (all cause mortality - mortality)*(Stock+Condition)/Condition.

The real dismod software actually takes in three of the following epidemiological
variables: incidence, remission, case fatality (or RR), prevalence, mortality, and generates the rest via the integrating system modeling.
}
\examples{
data('experiment1')
dismod_v2(pops=experiment1$population,
         alldeath_rate=experiment1$ACD_rate,
         condition_death_rate=experiment1$CD_rate,
         condition_incidence_rate=experiment1$CI_rate,
         duration=experiment1$CDuration,
         Stock=10000,
         Condition=5,
         start_age=0,
         age_range=c(0,1,seq(5,85,by=5)),
         end_age=85)         




}
\keyword{compartmental}
\keyword{dismod}
\keyword{framework,}
